feat: Add meta-tx to mocknet (#8641)
load_test_betanet and load_test_spoon_helper are 80% the same. Refactored them to a common file.
Changes account names in the mocknet test from "a00_something" to "a00.something".
Delete the unused skyward load test.
