+++ pytest/tests/mocknet/load_test_spoon.py
@@ -59,8 +59,6 @@ import pathlib
 
 sys.path.append(str(pathlib.Path(__file__).resolve().parents[2] / 'lib'))
 
-from helpers import load_test_spoon_helper
-from helpers import load_testing_add_and_delete_helper
 import mocknet
 import data
 
@@ -126,7 +124,8 @@ if __name__ == '__main__':
     parser.add_argument('--skip-setup', default=False, action='store_true')
     parser.add_argument('--skip-restart', default=False, action='store_true')
     parser.add_argument('--no-sharding', default=False, action='store_true')
-    parser.add_argument('--script', required=True)
+    # The flag is no longer needed but is kept for backwards-compatibility.
+    parser.add_argument('--script', required=False)
     parser.add_argument('--num-seats', type=int, required=True)
 
     args = parser.parse_args()
@@ -196,17 +195,12 @@ if __name__ == '__main__':
     logger.info(f'initial_validator_accounts: {initial_validator_accounts}')
     test_passed = True
 
-    script, deploy_time, test_timeout = (None, None, None)
-    if args.script == 'skyward':
-        script = 'load_testing_add_and_delete_helper.py'
-        deploy_time = load_testing_add_and_delete_helper.CONTRACT_DEPLOY_TIME
-        test_timeout = load_testing_add_and_delete_helper.TEST_TIMEOUT
-    elif args.script == 'add_and_delete':
-        script = 'load_test_spoon_helper.py'
-        deploy_time = load_test_spoon_helper.CONTRACT_DEPLOY_TIME
-        test_timeout = load_test_spoon_helper.TEST_TIMEOUT
-    else:
-        assert False, f'Unsupported --script={args.script}'
+    script = 'load_test_spoon_helper.py'
+    # TODO: Get these constants from load_test_spoon_helper:
+    # deploy_time = load_test_spoon_helper.CONTRACT_DEPLOY_TIME
+    # test_timeout = load_test_spoon_helper.TEST_TIMEOUT
+    deploy_time = 10 * mocknet.NUM_ACCOUNTS
+    test_timeout = 12 * 60 * 60
 
     if not args.skip_load:
         logger.info('Starting transaction spamming scripts.')
